+++ src/app/lib/auth.ts
@@ -1,7 +1,7 @@
 'use server';
 
 import { AuthError } from 'next-auth';
-import { signIn, auth } from '../../../auth';
+import { signIn, auth } from '../../auth';
 import { redirect } from 'next/navigation';
 
 export async function authenticate(
